Adds automated horizontal scrollbars to tables of times to handle overflow
Team EE has exposed that elmastats does not have handling for when tables of times exceed the expected maximum-width of the page. 

Though EE is currently the only example that is breaking, other teams are not far off and in theory we may choose to later add further functionality that does this too.

This commit adjusts the CSS of the table class across the board to automagically handle horizontal scrollbars rather than breaking.