Resolve CSS architecture issues (#3)
* Move scrollbar attempt to the correct css architecture

* Add automated horizontal scrollbar to div ID mid

* Remove random hi.css file

The current elmastats architecture is randomly loading hi.css from the main PHP directory rather than the hi.css which is specifically located in the CSS directory. This is extremely confusing and I am fixing it.

* Address random hi.css file issue

The current elmastats architecture is randomly loading hi.css from the main PHP directory rather than the hi.css which is specifically located in the CSS directory. This is extremely confusing and I am fixing it.

* Actually load CSS from the CSS directory

The current elmastats architecture is randomly loading css from the main PHP directory rather than the css which is specifically located in the CSS directory. This is extremely confusing and I am fixing it.

* Remove base directory CSS files

CSS should now be loaded from the CSS directory as intended, rather than the directory serving as a red herring for developers

* Reinstate and comment the target-time-colour handling